requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.4
et_xmlfile==2.0.0
idna==3.11
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.32.5
six==1.17.0
tzdata==2025.3
urllib3==2.6.3
